+++ TrollFools/InjectView.swift
@@ -26,6 +26,7 @@ struct InjectView: View {
     @AppStorage var useWeakReference: Bool
     @AppStorage var preferMainExecutable: Bool
     @AppStorage var useFrameworkEnumerationFallback: Bool
+    @AppStorage var deferPlugInLoading: Bool
     @AppStorage var injectStrategy: InjectorV3.Strategy
     @AppStorage var dryRun: Bool
 
@@ -35,6 +36,7 @@ struct InjectView: View {
         _useWeakReference = AppStorage(wrappedValue: true, "UseWeakReference-\(app.bid)")
         _preferMainExecutable = AppStorage(wrappedValue: false, "PreferMainExecutable-\(app.bid)")
         _useFrameworkEnumerationFallback = AppStorage(wrappedValue: true, "UseFrameworkEnumerationFallback-\(app.bid)")
+        _deferPlugInLoading = AppStorage(wrappedValue: false, "DeferPlugInLoading-\(app.bid)")
         _injectStrategy = AppStorage(wrappedValue: .lexicographic, "InjectStrategy-\(app.bid)")
         _dryRun = AppStorage(wrappedValue: false, "DryRun-\(app.bid)")
     }
@@ -140,6 +142,7 @@ struct InjectView: View {
             injector.useWeakReference = useWeakReference
             injector.preferMainExecutable = preferMainExecutable
             injector.useFrameworkEnumerationFallback = useFrameworkEnumerationFallback
+            injector.deferPlugInLoading = deferPlugInLoading
             injector.injectStrategy = injectStrategy
 
             if dryRun {